Add files via upload
In This project I have involved two major domains namely Web Application with AI (OpenAi) Api key to Act Like a chatbot to give the solution depending upon user's query,
with the help of java script make the app even more responsive. And also I have included the technologies of (Node.js, Express.js,) to run our our application in server to handle request and response of the web application. And I have Attached the Output screenshot in this respository.